refactor: controller package 생성후 ChessGame, menu 위치 변경
