+++ src/test/java/chess/view/MenuTest.java
@@ -2,7 +2,7 @@ package chess.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.menu.MenuType;
+import chess.controller.menu.MenuType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
